@@ -2,11 +2,11 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  before_filter :logged_in?
+  # before_filter :logged_in?
 
-  def logged_in?
-    unless User.find_by_id(session[:user_id])
-      redirect_to login_url, notice: "Please Log In!"
-    end
-  end
+  # def logged_in?
+  #   unless current_user
+  #     redirect_to new_user_session_path, notice: 'Please Log In!'
+  #   end
+  # end
 end
